+++ src/Models/Transaction/Flags/MPTokenAuthorizeFlags.php
@@ -15,7 +15,7 @@ use Hardcastle\XRPL_PHP\Models\Common\FlagSet;
 /**
  * Flags of the MPTokenAuthorize transaction.
  *
- * Values are those of rippled 3.3.0 (TxFlags.h); combine them with |.
+ * Values are those of rippled 3.4.0 (TxFlags.h); combine them with |.
  * https://xrpl.org/mptokenauthorize.html
  */
 final class MPTokenAuthorizeFlags